requirements.txt
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
keras==2.10.0
matplotlib==3.10.7
numpy==1.26.4
pandas==2.3.3
PyJWT==2.10.1
scikit-learn==1.7.2
tensorflow==2.10.0
yfinance==1.0